Fix typed proxy access to GenericSkeleton event storage
Typed ProxyEvent now accesses event samples through EventMetaInfo raw slot storage instead of interpreting provider storage as EventDataStorage<SampleType>.

This keeps typed proxies compatible with GenericSkeleton-created storage while preserving typed skeleton behavior. Proxy event tests and fixtures were updated to emulate production typed storage with a type-erased EventMetaInfo view. Existing proxy event construction tests were aligned with the fixture sample type so EventMetaInfo validation matches the dummy provider storage.

+++ score/mw/com/impl/bindings/lola/proxy_event_common_test.cpp
@@ -27,7 +27,9 @@ namespace score::mw::com::impl::lola
 namespace
 {
 
-using TestSampleType = std::uint16_t;
+// Keep this aligned with ProxyMockedMemoryFixture::SampleType because the fixture initializes
+// EventMetaInfo from its typed EventDataStorage backing storage.
+using TestSampleType = std::uint32_t;
 
 using namespace ::score::memory::shared;
 using ::testing::Return;
